# extracted_specs/openapi_golden_congno.yaml
openapi: 3.0.3
info:
  title: Cong No Golden API
  version: "1.0.0"
servers:
  - url: http://server1:8080/api
...
paths:
...
  /imports/upload:
    post:
      tags: [Imports]
      summary: Upload excel to staging
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, file]
              properties:
                type:
                  type: string
                  enum: [INVOICE, ADVANCE, RECEIPT]
                fromDate: { type: string, format: date }
                toDate: { type: string, format: date }
                file:
                  type: string
                  format: binary
      responses:
        "200": { description: OK }
  /imports/{batchId}/preview:
    get:
      tags: [Imports]
      summary: Preview staging rows with validations
      parameters:
        - in: path
          name: batchId
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /imports/{batchId}/commit:
    post:
      tags: [Imports]
      summary: Commit staging to main tables
      parameters:
        - in: path
          name: batchId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        "200": { description: OK }
...
  /advances/{id}/approve:
    post:
      tags: [Advances]
      summary: Approve advance
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /receipts/{id}/approve:
    post:
      tags: [Receipts]
      summary: Approve receipt
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...